Выносит hooks в отдельные файлы
