+++ package.json
@@ -10,31 +10,31 @@
     "test": "npm run eslint"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.13",
-    "@babel/preset-env": "^7.12.13",
-    "@babel/preset-react": "^7.12.13",
-    "babel-loader": "^8.2.2",
-    "css-loader": "^5.1.3",
+    "@babel/core": "7.12.13",
+    "@babel/preset-env": "7.12.13",
+    "@babel/preset-react": "7.12.13",
+    "babel-loader": "8.2.2",
+    "css-loader": "5.1.3",
     "eslint": "7.18.0",
     "eslint-config-htmlacademy": "2.1.0",
-    "eslint-plugin-react": "^7.22.0",
-    "style-loader": "^2.0.0",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.2"
+    "eslint-plugin-react": "7.22.0",
+    "style-loader": "2.0.0",
+    "webpack": "4.46.0",
+    "webpack-cli": "3.3.12",
+    "webpack-dev-server": "3.11.2"
   },
   "engines": {
     "node": ">=12.0.0"
   },
   "dependencies": {
-    "dayjs": "^1.10.4",
-    "history": "^4.10.1",
-    "prop-types": "^15.7.2",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-redux": "^7.2.2",
-    "react-router-dom": "^5.2.0",
-    "redux": "^4.0.5",
-    "redux-devtools-extension": "^2.13.8"
+    "dayjs": "1.10.4",
+    "history": "4.10.1",
+    "prop-types": "15.7.2",
+    "react": "16.14.0",
+    "react-dom": "16.14.0",
+    "react-redux": "7.2.2",
+    "react-router-dom": "5.2.0",
+    "redux": "4.0.5",
+    "redux-devtools-extension": "2.13.8"
   }
 }
