+++ src/components/promo-film/promo-film.jsx
@@ -1,18 +1,10 @@
 import React, {useEffect} from 'react';
-import {shallowEqual, useDispatch, useSelector} from 'react-redux';
+import {useDispatch} from 'react-redux';
 import Api from '../../api/api';
 import {getPromoFilm} from '../../store/action';
 import PromoFilmView from '../promo-film/promo-film-view';
 import LoadingScreen from '../loading-screen/loading-screen';
-
-// to storage ?
-const selectPromoFilm = (PROMO) => {
-  const found = PROMO.promoFilm;
-  return found;
-};
-
-const useSelectPromoFilm = () => useSelector(({PROMO}) => selectPromoFilm(PROMO), shallowEqual);
-
+import {useSelectPromoFilm} from '../../store/hooks/use-select-promo-film';
 
 const PromoFilm = () => {
   const api = new Api();
