+++ src/store/hooks/use-select-genres-array.js
@@ -1,0 +1,10 @@
+import {shallowEqual, useSelector} from "react-redux";
+
+const selectGenresArray = (FILMS) => {
+  const found = [...new Set(FILMS.films.map((film) => film.genre))];
+  return found;
+};
+
+const useSelectGenresArray = () => useSelector(({FILMS}) => selectGenresArray(FILMS), shallowEqual);
+
+export {selectGenresArray, useSelectGenresArray};
