+++ src/store/hooks/use-select-promo-film.js
@@ -1,0 +1,10 @@
+import {shallowEqual, useSelector} from "react-redux";
+
+const selectPromoFilm = (PROMO) => {
+  const found = PROMO.promoFilm;
+  return found;
+};
+
+const useSelectPromoFilm = () => useSelector(({PROMO}) => selectPromoFilm(PROMO), shallowEqual);
+
+export {selectPromoFilm, useSelectPromoFilm};
